% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_virtualenv.R
\name{create_virtualenv}
\alias{create_virtualenv}
\title{Create virtual environment}
\usage{
create_virtualenv()
}
\description{
Creates a Python virtual environment and installs necessary packages
}
\details{
This function creates a Python virtual environment called "HIP_env" and installs
necessary packages such as torch, numpy, and matplotlib into the environment. Once
created, the environment will be used when the package is loaded.
}
\examples{
create_virtualenv()

}
