% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate simulated data to use with HIP}
\usage{
generate_data(seed=1, n=c(250,260), p=c(300,350), K=2, D=2, S=2, nonzero=50, offset=25,
sigma_x=1, sigma_y=1, family='gaussian', test_data=FALSE,
q=1, m = 2, theta_init=NULL, B=NULL, Z=NULL, y_mean=0,
y_sd=1, z_mean=25, z_sd=3, beta=0, tau=NULL, t_dist=FALSE, t_df=20)
}
\arguments{
\item{seed}{int or string - set a seed for replicability}

\item{n}{int vector - number of subjects desired in each subgroup; should be length S}

\item{p}{int vector - number of covariates desired in each data set; should be length D}

\item{K}{int - number of latent components to use in generating the data}

\item{D}{int - number of data sets to generate}

\item{S}{int - number of subgroups}

\item{nonzero}{int - number of important variables for each subgroup; same for all subgroups}

\item{offset}{int - how many variables to offset the important variables between subgroups; same for all subgroups}

\item{sigma_x}{double - factor by which the errors added to X are scaled by}

\item{sigma_y}{double - factor by which the errors added to Y are scaled by; not applicable when family = 'poisson' or 'zip'}

\item{family}{string - determines what type of outcome is generated:
\itemize{
\item 'gaussian' - q should also be specified or q will default to 1
\item 'multiclass' - m should also be specified or m will default to 2
\item 'poisson'
\item 'zip'
}}

\item{test_data}{bool - TRUE to generate test data, FALSE otherwise}

\item{q}{int - number of continuous outcomes; can be >= 1; default = 1}

\item{m}{int - number of classes in multiclass outcome; can be >= 2; default = 2}

\item{theta_init}{double matrix - matrix to use for theta; if None, then a theta matrix will be generated from a U(0,1) distribution}

\item{B}{matrix list - B^d,s matrices used to generate the data; will randomly generate if not provided}

\item{Z}{matrix list - Z^s matrices used to generate the data; if not provided, will generate as N(mu = z_mean, sigma = z_sd)}

\item{y_mean}{double - Value for mean of Y; default = 0}

\item{y_sd}{double - Value for sd of Y; default = 1}

\item{z_mean}{double - Value for mean of Y; default = 0}

\item{z_sd}{double - Value for sd of Y; default = 1}

\item{beta}{double - Value of intercept term; default = 0}

\item{t_dist}{boolean - if family = "gaussian", setting t_dist = TRUE will generate data with errors drawn from a t-distribution with some specified degrees of freedom}

\item{t_df}{int - degrees of freedom for t-distribution if t_dist = TRUE; default = 20}
}
\value{
\code{generate_data} returns a nested list object with the following elements:\tabular{ll}{
   \code{X} \tab matrix list - Contains X^d,s matrices; access as \code{data$X[[d]][[s]]} \cr
   \code{Y} \tab matrix list - Contains Y^s matrices; access as \code{data$Y[[d]][[s]]} \cr
   \code{X_test} \tab matrix list - Contains test data X^d,s matrices; access as \code{data$X_test[[d]][[s]]} \cr
   \code{Y_test} \tab matrix list - Contains test data Y^s matrices; access as \code{data$Y_test[[d]][[s]]} \cr
   \code{D} \tab int - Number of data sets generated \cr
   \code{S} \tab int - Number of subgroups \cr
   \code{sub_vec} \tab int vec - Vector with labels for subgroups (from 1 to S) \cr
   \code{Z} \tab matrix list - Contains all Z^s matrices used to generate the data; access as \code{data[[Z]][[d]][[s]]} \cr
   \code{B} \tab matrix list - Contains all B^d,s matrices used to generate the data; access as \code{data[[B]][[d]][[s]]} \cr
   \code{theta} \tab matrix - Contains the theta matrix used in generating the data; access as \code{data$theta} \cr
   \code{beta} \tab double - Value for intercept \cr
   \code{tau} \tab double - If family = 'zip', proportion of observations in zero state with default = 0.25; None ow \cr
   \code{y_mean} \tab double - Value for mean of Y \cr
   \code{y_sd} \tab double - Value for sd of Y \cr
   \code{z_mean} \tab double - Value for mean of Z \cr
   \code{z_sd} \tab double - Value for sd of Z \cr
   \code{seed} \tab int or string - Value of seed used to generate the data \cr
   \code{var_list} \tab vec list - Contains D vectors with labels for variables in each dataset (1 to p^d) \cr
}
}
\description{
\code{generate_data} returns a nested list containing data matrices and values used to generate the data.
}
\details{
\code{generate_data} can be used to generate simulated multi-view data across different subgroups. The input
parameters specify relevant information such as the number of data views,
the number of subgroups, subjects desired for each subgroup, and more. The output of \code{generate_data}
gives the \eqn{X^{d,s}} data matrices and \eqn{Y^s} outcome matrices, as well as values used to
generate the data such as the \eqn{Z^s} and \eqn{B^{d,s}} matrices. See the HIP paper for details.
}
\examples{
# The values below are the function defaults, but can be changed for other simulated datasets
family <- 'gaussian'
seed <- 1
nonzero <- 50
offset <- 25
n <- c(250,260)
K <- 2
S <- 2
D <- 2
p <- c(300,350)
sigma_x <- 1.0
sigma_y <- 1.0

# Generating data (no test data)
data_gen <- generate_data(seed, n, p, K, D, S, nonzero, offset, sigma_x,
sigma_y, family)

# Generating data (with test data)
data_gen_test <- generate_data(seed, n, p, K, D, S, nonzero, offset,
                              sigma_x, sigma_y, family, test_data=TRUE)

}
