% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIP-package.R
\docType{package}
\name{HIP-package}
\alias{HIP}
\alias{HIP-package}
\title{HIP: Heterogeneity in Integration and Prediction (HIP)}
\description{
We present HIP (Heterogeneity in Integration and Prediction) for integration of multiple data views while accounting for subgroup heterogeneity to identify common and subgroup-specific variables of interest for a particular outcome. HIP accommodates Gaussian, multi-class, Poisson, and zero-inflated Poisson outcomes. In addition to this package, HIP may be accessed through an R Shiny application found at \url{https://multi-viewlearn.shinyapps.io/HIP_ShinyApp/}. For details, refer to the paper found on arXiv: \url{https://arxiv.org/abs/2111.06962}
}
\keyword{internal}
