% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIP_pred.R
\name{HIP_pred}
\alias{HIP_pred}
\title{Make predictions from HIP model}
\usage{
HIP_pred(X_dat, res, fix_or_search, standardize="no")
}
\arguments{
\item{X_dat}{matirx list - X^d,s matrices from \code{generate_data} or \code{format_data}}

\item{res}{output from \code{select_lambda} or \code{fixed_lambda}}

\item{fix_or_search}{string - 'fixed' if \code{res} comes from \code{fixed_lambda}, else 'search'}

\item{standardize}{string - 'yes' to standardize Y data if it was standardized in \code{select_lambda}, else
return unstandardized predicted Y values}
}
\value{
The function returns a list of \eqn{S} tensors containing predictions for
observations in each subgroup.
}
\description{
Outputs predicted Y values given X data and results from \code{select_lambda} or \code{fixed_lambda}.
}
\examples{
# Generate data
dat_gen_test <- generate_data(test_data=TRUE)

# Get results from select_lambda
res <- select_lambda(dat_gen_test$X, dat_gen_test$Y, c(1,1), 'gaussian', 50,
                     K = 2, num_steps=c(4,4))

# Make predictions on X test data (and standardize predictions)
Y_pred <- HIP_pred(dat_gen_test$X_test, res, 'search', 'yes')

}
