% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_dat.R
\name{standardize_dat}
\alias{standardize_dat}
\title{Standardizes input data}
\usage{
standardize_dat(
  standardize,
  std_type,
  X = NULL,
  Y = NULL,
  X_train = NULL,
  Y_train = NULL,
  std_x = TRUE,
  std_y = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{standardize}{string - One of "all", "subgroup", or "none"}

\item{std_type}{string - One of "scale_center", "center", or "norm"}

\item{X}{matrix list - list of X^d,s matrices containing covariates}

\item{Y}{matrix list - list of Y^s matrices containing outcomes}

\item{X_train}{matrix list - list of X^d,s matrices whose mean and sd will be used to standardize X}

\item{Y_train}{matrix list - list of Y^s matrices whose mean andd sd will be used to standardize Y}

\item{std_x}{boolean - indicates whether to standardize X; default = True}

\item{std_y}{boolean - indicates whether to standardize Y; default = True}
}
\value{
Returns a nested list with the following elements:\tabular{ll}{
   \code{X} \tab matrix list - standardized X^d,s if standardization was requested \cr
   \code{Y} \tab matrix list - standardized Y^s if standardization was requested \cr
}
}
\description{
Helper function used to standardize testing data from training data.
}
